Full connection to all APIs
Frontend takes in user input of song and artist name. Frontend sends info to the backend, where it is processed and an API request is made to Spotify. Backend then connects to Genius API to return lyrics of the song. These lyrics are then sent to Watson API to return the sentiment analysis.